sql: move coercion and casting to new typeconv module
Originally suggested by @sploiselle. The typeconv module subsumes the
existing cast module and incorporates more of the casting and coercion
logic.
